Move ACS tests to its own file
